Add relevant platforms as child list of destination
Create a string of destinations to append to list rather than doing
too much DOM manipulation; do it in one go for performance.
Walk through the JSON data to see which platforms any given destination
can be reached from and create a list which is appended to within the
destination list item.