Break loop up with setTimeout and add Platforms view
Because the amount of data being processed is large, split each iteration
with a timeout to allow the browser to still be responsive.
Add an overlay to stop users clicking around the interface while processing
is in progress.
Check typeof callback is function before calling
Only show the overlay if there are more than 10 itterations to perform.
Add functions to display a list of platforms and all the destinations they serve
Hook up links to switch the view from destinations to platforms and vice versa
Comment out GA code until I add the correct ID.